requirements.txt
Django==2.2.2
djangorestframework==3.9.4
pip==19.0.3
python-decouple==3.1
pytz==2019.1
setuptools==40.8.0
sqlparse==0.3.0